-include the c wrapper for LuP -change some stupid things for Mirosoft Visual Studio (sprintf() is deprecated) -include transform plane -change something in structure Applanix->transform
git-svn-id: http://www/repos/lib/photogrammetrie/trunk@973 9521289c-3fb7-4341-aae7-02de4671f267

+++ wrapper_for_java/CBPointList.cpp
@@ -14,7 +14,7 @@ void CBPointList::addBPoint(BPoint bp)
 }
 int CBPointList::size()
 {
-	return m_BPoint_List.size();
+	return static_cast<int>(m_BPoint_List.size());
 }
 void CBPointList::clearBPointList()
 {
